+++ rt-src/rt/cwind_builtin.c
@@ -196,9 +196,11 @@ static bool cwfmt_record(CwFmtCtx_t* c, const CWindObject_t* obj) {
     case CWByte:
         return cwfmt_printf(c, "%u", *(const uint8_t*)(uintptr_t)h->address);
     case CWInt32:
-        return cwfmt_printf(c, "%d", *(const int32_t*)(uintptr_t)h->address);
+        return cwfmt_printf(c, "%lld",
+                            (long long)*(const int32_t*)(uintptr_t)h->address);
     case CWUInt32:
-        return cwfmt_printf(c, "%u", *(const uint32_t*)(uintptr_t)h->address);
+        return cwfmt_printf(c, "%llu",
+                            (unsigned long long)*(const uint32_t*)(uintptr_t)h->address);
     case CWInt64:
         return cwfmt_printf(c, "%lld",
                             (long long)*(const int64_t*)(uintptr_t)h->address);
